requirements.txt
pyautogui
pytesseract
pillow
keyboard
flask
requests
python-dotenv
flask-cors
